+++ src/features/BrowseBiomodels.ts
@@ -50,3 +50,47 @@ export async function getModel(modelId: string) {
         throw error;
     }
 }
+
+export function getBiomodels(setLoading: React.Dispatch<React.SetStateAction<boolean>>, setChosenModel: React.Dispatch<React.SetStateAction<string | null>>) {
+  const biomodelBrowse = document.getElementById('biomodel-browse') as HTMLInputElement;
+  const dropdown = document.getElementById('dropdown');
+  var biomodels: any;
+  var chosenModel: any;
+
+  biomodelBrowse.addEventListener('keyup', async (val) => {
+    const biomodel = val;
+    if ((val.target as HTMLInputElement).value.length < 3) {
+      dropdown!.innerHTML = "";
+      return;
+    }
+    setTimeout(async () => {
+      setLoading(true);
+      dropdown!.innerHTML = "";
+      biomodels = await searchModels(biomodel);
+      // If no models found, display "No models found"
+      if (biomodels.models.size === 0) {
+        setLoading(false);
+        biomodels = null;
+        const li = document.createElement('li');
+        li.innerHTML = "No models found";
+        dropdown!.innerHTML = "";
+        dropdown!.appendChild(li);
+        return;
+      }
+      // Otherwise, display the models in the dropdown
+      dropdown!.style.display = "block";
+      biomodels.models.forEach(function (model: any) {
+        setLoading(false);
+        const a = document.createElement('a');
+        a.addEventListener('click', () => {
+          biomodelBrowse.value = "";
+          dropdown!.innerHTML = "";
+          chosenModel = model.id;
+          setChosenModel(chosenModel);
+        });
+        a.innerHTML = model.name + ": " + model.id + "\n";
+        dropdown!.appendChild(a);
+      });
+    }, 300);
+  });
+}
